feature: NSO XE to Openconfig system translation script is now a package for better organization
