+++ test/tests_nso_to_oc/xe_system.yml
@@ -309,7 +309,7 @@
               last: 4
         api_method: PATCH
     - name: Execute NSO NED to OC script
-      script: ../tools/translate_nso_device_config_to_openconfig.py
+      script: ../../package_nso_to_oc/xe/xe_system.py
       args:
         executable: python3
       register: nso_to_oc_result
@@ -410,9 +410,8 @@
             key-id: 1
             key-type: openconfig-system:NTP_AUTH_MD5
             key-value: password1
-
     - name: Execute NSO NED to OC script
-      script: ../tools/translate_nso_device_config_to_openconfig.py
+      script: ../../package_nso_to_oc/xe/xe_system.py
       args:
         executable: python3
       register: nso_to_oc_result
